// data/upgrades/enums.ts
export enum UpgradeId {
  BASIC_BED = "BASIC_BED",
  STRAW_BED = "STRAW_BED",
  COMFORTABLE_BED = "COMFORTABLE_BED",
  REFINED_BED = "REFINED_BED",
  LUXURY_BED = "LUXURY_BED",

  BASIC_HAMMER = "BASIC_HAMMER",
  BRONZE_HAMMER = "BRONZE_HAMMER",
  BRONZE_HAMMER1 = "BRONZE_HAMMER1",
  BRONZE_HAMMER2 = "BRONZE_HAMMER2",
  BRONZE_HAMMER3 = "BRONZE_HAMMER3",
  BRONZE_HAMMER4 = "BRONZE_HAMMER4",
  IRON_HAMMER = "IRON_HAMMER",
  IRON_HAMMER1 = "IRON_HAMMER1",
  IRON_HAMMER2 = "IRON_HAMMER2",
  IRON_HAMMER3 = "IRON_HAMMER3",
  IRON_HAMMER4 = "IRON_HAMMER4",
  STEEL_HAMMER = "STEEL_HAMMER",
  STEEL_HAMMER1 = "STEEL_HAMMER1",
  STEEL_HAMMER2 = "STEEL_HAMMER2",
  STEEL_HAMMER3 = "STEEL_HAMMER3",
  STEEL_HAMMER4 = "STEEL_HAMMER4",
  MITHRIL_HAMMER = "MITHRIL_HAMMER",
  MITHRIL_HAMMER1 = "MITHRIL_HAMMER1",
  MITHRIL_HAMMER2 = "MITHRIL_HAMMER2",
  MITHRIL_HAMMER3 = "MITHRIL_HAMMER3",
  MITHRIL_HAMMER4 = "MITHRIL_HAMMER4",
  ADAMANT_HAMMER = "ADAMANT_HAMMER",
  ADAMANT_HAMMER1 = "ADAMANT_HAMMER1",
  ADAMANT_HAMMER2 = "ADAMANT_HAMMER2",
  ADAMANT_HAMMER3 = "ADAMANT_HAMMER3",
  ADAMANT_HAMMER4 = "ADAMANT_HAMMER4",

  BASIC_PICKAXE = "BASIC_PICKAXE",
  BRONZE_PICKAXE = "BRONZE_PICKAXE",
  BRONZE_PICKAXE1 = "BRONZE_PICKAXE1",
  BRONZE_PICKAXE2 = "BRONZE_PICKAXE2",
  BRONZE_PICKAXE3 = "BRONZE_PICKAXE3",
  BRONZE_PICKAXE4 = "BRONZE_PICKAXE4",
  IRON_PICKAXE = "IRON_PICKAXE",
  IRON_PICKAXE1 = "IRON_PICKAXE1",
  IRON_PICKAXE2 = "IRON_PICKAXE2",
  IRON_PICKAXE3 = "IRON_PICKAXE3",
  IRON_PICKAXE4 = "IRON_PICKAXE4",
  STEEL_PICKAXE = "STEEL_PICKAXE",
  STEEL_PICKAXE1 = "STEEL_PICKAXE1",
  STEEL_PICKAXE2 = "STEEL_PICKAXE2",
  STEEL_PICKAXE3 = "STEEL_PICKAXE3",
  STEEL_PICKAXE4 = "STEEL_PICKAXE4",
  MITHRIL_PICKAXE = "MITHRIL_PICKAXE",
  MITHRIL_PICKAXE1 = "MITHRIL_PICKAXE1",
  MITHRIL_PICKAXE2 = "MITHRIL_PICKAXE2",
  MITHRIL_PICKAXE3 = "MITHRIL_PICKAXE3",
  MITHRIL_PICKAXE4 = "MITHRIL_PICKAXE4",
  ADAMANT_PICKAXE = "ADAMANT_PICKAXE",
  ADAMANT_PICKAXE1 = "ADAMANT_PICKAXE1",
  ADAMANT_PICKAXE2 = "ADAMANT_PICKAXE2",
  ADAMANT_PICKAXE3 = "ADAMANT_PICKAXE3",
  ADAMANT_PICKAXE4 = "ADAMANT_PICKAXE4",

  BASIC_FORGE = "BASIC_FORGE",
  BASIC_ANVIL = "BASIC_ANVIL",
  COAL_MINE_CANERY = "COAL_MINE_CANERY",
}